+++ src/lib/string.c
@@ -1,7 +1,8 @@
 #include <string.h>
+#include <stdint.h>
 #include <types.h>
 
-void *memset(void* dest, register int value, register int len) {
+void *memset(void* dest, register int value, register uint64_t len) {
 	register unsigned char *ptr = (unsigned char*)dest;
 	while(len-- > 0) {
 		*ptr++ = value;
@@ -9,7 +10,7 @@ void *memset(void* dest, register int value, register int len) {
 	return dest;
 }
 
-void *memcpy(void *dest, void *src, register int len) {
+void *memcpy(void *dest, void *src, register uint64_t len) {
 	register unsigned char *bdest = (unsigned char*)dest;
 	register unsigned char *bsrc = (unsigned char*)src;
 	while(len-- > 0) {
